Add elasticsearch struct
