+++ cli/Cargo.toml
@@ -16,9 +16,11 @@ ascii_table = "3.0"
 base64 = "0.13"
 chrono = "0.4"
 clap = "2.33.1"
+elasticsearch = "7.11.0-alpha.1"
 hyper = { version = "0.14", features = ["client", "http1", "http2", "runtime", "server"] }
 hyper-tls = "0.5" # needed for HTTPS w/ hyper
-log = { version = "0.4.14", features = ["std"] }
+log = { version = "0.4.14", features = ["serde", "std"] }
+once_cell = "1.5.2"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = { version = "1.0" }
 spinner = "0.5.0"
